typos: fix unexpected behaviour
Use excludes from .typos.toml when running `pre-commit run typos --all-files`.

Since we don't set "pass_filenames = false" for typos (see
upstream discussions [0]), we must ensure that pre-commit never
passes the files specified as excludes in `.typos.toml` to
`typos` as argument, if possible. Otherwise,
`$ pre-commit run typos --all-files` has another behaviour than
`$ typos .`, which is confusing and wrong.

To not break anything, and to be compliant with the existing
API, we encourage users to provide the `.typos.toml` as Nix
path, so the excludes can be read via evaluation time.

The `configAsFile` variable is the path where the configuration
file can be found by the `typos` utility when it executes in the
user environment. Not necessarily in Nix store.

[0]: https://github.com/cachix/pre-commit-hooks.nix/pull/387#issuecomment-1893600631